backend/requirements.txt
# Pravaah OS — Backend Python Dependencies (v2)
# Install with: pip install -r backend/requirements.txt

# Flask core
flask==3.0.3
flask-cors==4.0.1
flask-socketio==5.3.6

# Async mode
gevent==24.2.1
gevent-websocket==0.10.1

# WebSocket async support for SocketIO (fallback/threading mode)
simple-websocket==1.0.0

# HTTP clients (TTS + OpenRouter)
requests==2.32.3

# Environment variable loading
python-dotenv==1.0.1

# File validation
werkzeug==3.0.3

# PDF text extraction
pdfminer.six==20231228

# DOCX text extraction
python-docx==1.1.2

# Typing support
typing-extensions==4.12.2

# SQLAlchemy ORM (v2)
sqlalchemy==2.0.30

# Redis (v2)
redis==5.0.7

# RQ job queue (v2)
rq==1.16.2
rq-dashboard==0.6.1

# JWT authentication (v2)
PyJWT==2.8.0

# Password hashing (v2)
bcrypt==4.1.3

# Rate limiting (v2)
flask-limiter==3.7.0

# Structured logging (v2)
structlog==24.1.0

# Vector store (v2)
chromadb==0.5.0

# Token counting (v2)
tiktoken==0.7.0

# PostgreSQL driver (required for Supabase)
psycopg2-binary==2.9.9

# Supabase Python client
supabase==2.4.6

# Deepgram STT WebSocket
websockets==12.0
